requirements.txt
# Core AI/ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=2.0.0
transformers>=4.21.0
langchain>=0.1.0
openai>=1.0.0

# Data Processing
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
python-dotenv>=0.19.0

# Newsletter & Content Creation
jinja2>=3.1.0
markdown>=3.4.0
python-slugify>=6.1.0
pillow>=9.0.0

# Web Scraping & APIs
selenium>=4.8.0
tweepy>=4.12.0
praw>=7.6.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.12.0

# Vector Databases & Search
chromadb>=0.4.0
faiss-cpu>=1.7.0
pinecone-client>=2.2.0

# MLOps & Monitoring
mlflow>=2.0.0
wandb>=0.13.0

# Utilities
tqdm>=4.64.0
click>=8.1.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=6.4.0